prserver-main/requirements.txt
asgiref==3.8.1
attrs==24.1.0
awscli==1.40.2
bcrypt==4.2.1
blinker==1.9
boto3==1.38.3
botocore==1.38.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.8
django-allauth==0.57.2
django-allauth-2fa==0.12.0
django-background-tasks==1.2.8
django-cors-headers
django-csp
django-formtools==2.5.1
django-ipware
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-ratelimit
django-ses
django-two-factor-auth==1.17.0
docutils==0.19.0
filelock==3.18.0
Flask==3.1.0
greenlet==3.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==2.1.5
msgpack==1.0.5
ntlm-auth==1.5.0
oauthlib==3.2.2
packaging==25.0
paramiko==3.5.0
phonenumbers==9.0.3
psutil==7.0.0
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.22.0
pygdbmi==0.11.0.0
pypng==0.20220715.0
pyspnego==0.10.2
python-dateutil==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.1
pywinrm==0.5.0
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
rsa==4.7
s3transfer==0.12.0
sentry-sdk==2.27.0
six==1.17.0
sqlparse==0.5.1
stripe
typing_extensions==4.12.2
urllib3==2.4.0
web.py
Werkzeug==3.1.3
wheel==0.45.1
xmltodict==0.14.2
concurrently==2.1
